refactor(api): modularisation des sources
