+++ app/services/sources/index.ts
@@ -1,20 +1,24 @@
 import { Source } from '@/app/types/source';
 import toomicsSource from './toomics';
+import webtoonsSource from './webtoons';
+import komgaSource from './komga';
+import mangadexSource from './mangadex';
+import mangaScantradSource from './mangaScantrad';
 
-// Exportation de toutes les sources disponibles
 export const sources: Record<string, Source> = {
   toomics: toomicsSource,
+  webtoons: webtoonsSource,
+  komga: komgaSource,
+  mangadex: mangadexSource,
+  mangascantrad: mangaScantradSource,
 };
 
-// Exportation individuelle des sources
-export { toomicsSource };
+export { toomicsSource, webtoonsSource, komgaSource, mangadexSource, mangaScantradSource };
 
-// Fonction utilitaire pour obtenir toutes les sources
 export function getAllSources(): Source[] {
   return Object.values(sources);
 }
 
-// Fonction utilitaire pour obtenir une source par son nom
 export function getSource(name: string): Source | undefined {
   return sources[name];
 }
